Don't refresh the change log when clicking the "Download report as pdf" button. Fixes #1015.
