+++ components/frontend/src/utils.js
@@ -1,3 +1,6 @@
+import React from 'react';
+import { toast } from 'react-semantic-toasts';
+
 export function get_metric_name(metric, datamodel) {
     return metric.name || datamodel.metrics[metric.type].name;
 }
@@ -13,3 +16,14 @@ export function get_subject_name(subject, datamodel) {
 export function get_metric_target(metric) {
     return (metric.accept_debt ? metric.debt_target : metric.target) || "0";
 }
+
+export function show_message(type, title, description, icon) {
+    toast({
+      title: title,
+      type: type,
+      icon: icon,
+      size: "large",
+      description: <p>{description}</p>,
+      time: 30000
+    }, () => { }, () => { }, () => { });  // Event handlers are mandatory
+  }
